monai_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.3.0=pypi_0
alembic=1.8.1=pypi_0
asttokens=2.0.8=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2022.07.19=h06a4308_0
cachetools=5.2.0=pypi_0
certifi=2022.9.24=py310h06a4308_0
charset-normalizer=2.1.1=pypi_0
cinemagoer=2022.2.11=pypi_0
click=8.1.3=pypi_0
cloudpickle=2.2.0=pypi_0
connected-components-3d=3.10.5=pypi_0
contourpy=1.0.5=pypi_0
cycler=0.11.0=pypi_0
databricks-cli=0.17.3=pypi_0
debugpy=1.6.3=pypi_0
decorator=5.1.1=pypi_0
docker=6.0.0=pypi_0
docopt=0.6.2=pypi_0
einops=0.5.0=pypi_0
entrypoints=0.4=pypi_0
et-xmlfile=1.1.0=pypi_0
executing=1.1.1=pypi_0
filelock=3.8.0=pypi_0
flask=2.2.2=pypi_0
fonttools=4.37.4=pypi_0
gdown=4.5.1=pypi_0
gitdb=4.0.9=pypi_0
gitpython=3.1.29=pypi_0
google-auth=2.12.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
greenlet=1.1.3.post0=pypi_0
grpcio=1.49.1=pypi_0
gunicorn=20.1.0=pypi_0
huggingface-hub=0.10.1=pypi_0
idna=3.4=pypi_0
imageio=2.22.1=pypi_0
imdbpy=2022.7.9=pypi_0
importlib-metadata=4.13.0=pypi_0
install=1.3.5=pypi_0
ipykernel=6.16.0=pypi_0
ipython=8.5.0=pypi_0
itsdangerous=2.1.2=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jupyter-client=7.4.2=pypi_0
jupyter-core=4.11.1=pypi_0
kiwisolver=1.4.4=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.0.3=h7f8727e_2
lmdb=1.3.0=pypi_0
lxml=4.9.1=pypi_0
mako=1.2.3=pypi_0
markdown=3.4.1=pypi_0
markupsafe=2.1.1=pypi_0
matplotlib=3.6.1=pypi_0
matplotlib-inline=0.1.6=pypi_0
mlflow=1.29.0=pypi_0
monai=1.1.0+77.gf5708ea8=pypi_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.6=pypi_0
networkx=2.8.7=pypi_0
nibabel=4.0.2=pypi_0
nptyping=2.3.1=pypi_0
numpy=1.23.4=pypi_0
oauthlib=3.2.1=pypi_0
openpyxl=3.1.2=pypi_0
openssl=1.1.1q=h7f8727e_0
packaging=21.3=pypi_0
pandas=1.5.0=pypi_0
parso=0.8.3=pypi_0
patsy=0.5.3=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.2.0=pypi_0
pip=22.2.2=py310h06a4308_0
prometheus-client=0.15.0=pypi_0
prometheus-flask-exporter=0.20.3=pypi_0
prompt-toolkit=3.0.31=pypi_0
protobuf=3.19.6=pypi_0
psutil=5.9.2=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pydicom=2.3.1=pypi_0
pygments=2.13.0=pypi_0
pyjwt=2.5.0=pypi_0
pykwalify=1.8.0=pypi_0
pynrrd=1.0.0=pypi_0
pyparsing=3.0.9=pypi_0
pyradiomics=3.0.1=pypi_0
pysocks=1.7.1=pypi_0
python=3.10.6=haa1d7c7_0
python-dateutil=2.8.2=pypi_0
pytorch-ignite=0.4.10=pypi_0
pytz=2022.4=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0=pypi_0
pyzmq=24.0.1=pypi_0
querystring-parser=1.2.4=pypi_0
readline=8.1.2=h7f8727e_1
regex=2022.9.13=pypi_0
requests=2.28.1=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
ruamel-yaml=0.17.32=pypi_0
ruamel-yaml-clib=0.2.7=pypi_0
scikit-image=0.19.3=pypi_0
scikit-learn=1.0.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=63.4.1=py310h06a4308_0
simpleitk=2.2.1=pypi_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlalchemy=1.4.41=pypi_0
sqlite=3.39.3=h5082296_0
sqlparse=0.4.3=pypi_0
stack-data=0.5.1=pypi_0
statsmodels=0.14.0=pypi_0
tableone=0.8.0=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.10.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2022.10.10=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.1=pypi_0
torch=1.12.1+cu113=pypi_0
torchaudio=0.12.1+cu113=pypi_0
torchvision=0.13.1+cu113=pypi_0
tornado=6.2=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.4.0=pypi_0
transformers=4.23.1=pypi_0
typing-extensions=4.4.0=pypi_0
tzdata=2022c=h04d1e81_0
urllib3=1.26.12=pypi_0
wcwidth=0.2.5=pypi_0
websocket-client=1.4.1=pypi_0
werkzeug=2.2.2=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xgboost=1.6.2=pypi_0
xz=5.2.6=h5eee18b_0
zipp=3.9.0=pypi_0
zlib=1.2.12=h5eee18b_3
